fix interpolation error between years if target year is outside range of database years
If the target year is outside of range of database years, the interpolation has previously not interpolated correctly. It would give negative weights for one year and positive weights for another, while we would like to have a weight of 1 for the closest database year available. The new code is also a bit more simple.